==> notebooks/4_xgboost.ipynb <==
import pickle
import numpy as np

import xgboost
# from sklearn.metrics import accuracy_score
# from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split

with open('reviews_post_feature_eng.pkl', 'rb') as f:
    reviews = pickle.load(f)

#### Feature Importance: All Features

def select_fetures(dataset, features):
    return dataset[features]

#### Feature Importance: Selected Features

features = ['HelpfullnessRank', 'Score', 'Time', 'WordsCount', 'Total_Reviews_by_Reviewer',
        'ProductFrequency', 'WordCount', 'WordCountSummary', 'StopWords',
        'UpperCount', 'LowerCount', 'LowerCountSummary', 'DotCount',
        'CountPunctuation', 'CountDigits', 'Lexical', 'UpperLowerR',
        'UpperLowerSumR', 'DotCapitalR', 'DotCapitalSumR', 'CapitalsRatio',
        'neg', 'neu', 'pos', 'compound', 'ProductFreqlog',
        'ReviewsbyReviewerlog', 'WordCountlog', 'Month', 'Day']

reviews_features = select_fetures(reviews, features)

label = 'HelpfullnessRank'
X = reviews_features.drop(label, axis=1)
y = reviews_features[label]

# Split to train, test and validation datasets:
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.4)
X_test, X_val, y_test, y_val = train_test_split(X_test, y_test, test_size=0.5)
X_real_test = X_test
y_test_real = y_test
X_train = X_train
y_train = y_train
X_test = X_val
y_test = y_val

# Note

from numpy import loadtxt
from xgboost import XGBClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score

model = XGBClassifier(learning_rate=0.5, max_depth=10, n_estimators=200)
model.fit(X_train, y_train)

# make predictions for test data
y_pred = model.predict(X_test)
predictions = [round(value) for value in y_pred]

# evaluate predictions
accuracy = accuracy_score(y_test, predictions)
print("Accuracy: %.2f%%" % (accuracy * 100.0))

y_pred_train = model.predict(X_train)
predictions_train = [round(value) for value in y_pred_train]

accuracy = accuracy_score(y_train, predictions_train)
print("Accuracy: %.2f%%" % (accuracy * 100.0))

y_pred_test = model.predict(X_real_test)
predictions_test = [round(value) for value in y_pred_test]

accuracy = accuracy_score(y_test_real, predictions_test)
print("Accuracy: %.2f%%" % (accuracy * 100.0))

train_prob = model.predict_proba(X_train)
val_prob = model.predict_proba(X_test)
test_prob = model.predict_proba(X_real_test)

np.save('test_prob', test_prob)
np.save('val_prob', val_prob)
np.save('train_prob', train_prob)

3
