==> notebooks/9_ API_and_printing.ipynb <==
def rank_products_reviews(reviews, labels,  model, text_features, product_id):
    """
    Rank the reviews of a Product by their helpfullness given a product id and a model

    Args:
        reviews(pd.DataFrame): product reviews data-set (validation)
        labels(np.array): the labels of product's helpfullness
        model: the model used for classifying the reviews
        text_feature: the model which is used to extract features from the textual review
        product_id(str): the product id

    Returns:
        pd.DataFrame: Ranked Reviews

    """
    # Query all the reviews of product_id:
    products_reviews = reviews[reviews.ProductId == product_id]
    product_labels = labels[labels.ProductId == product_id]['HelpfullnessRank']

    # Extract the textual features:
    products_text_features = text_features.transform(products_reviews['Text'])

    # Predict the helpfullness score of the reviews:
    rnd_producy_pred = model.predict(products_text_features)

    products_reviews['Review_Rank'] = pd.Series(rnd_producy_pred, index=products_reviews.index)
    products_reviews['Original_Rank'] = pd.Series(product_labels, index=products_reviews.index)
    products_reviews = products_reviews.sort_values(
        by=['Review_Rank', 'Score', 'Total_Reviews_by_Reviewer', 'WordsCount'], ascending=False)
    

    ranked_reviews = products_reviews.drop(['Id', 'ProductId', 'UserId', 'Time', 'ProfileName',
                                            'HelpfulnessNumerator', 'HelpfulnessDenominator', 'ProductFrequency'],
                                           axis=1)

    return ranked_reviews


def print_reviews_by_ranking(df):
    """
    """
    for i in range(len(df)):
        print("my Rank: {}  Groud Trouth Rank: {}".format(df['Review_Rank'].iloc[i], df['Original_Rank'].iloc[i]))
        print("Text: ", df['Text'].iloc[i])
        print("Summary:", df['Summary'].iloc[i])
        print("\n")


def print_products_reviews(X_val, y_val,  clf, text_features, product_id):
    """
    """
    a= rank_products_reviews(X_val, y_val,  clf, text_features, product_id)
    print_reviews_by_ranking(a)

product_ids = X_val.ProductId.values
ranked_results = rank_products_reviews(X_val, y_val,  clf, text_features, product_ids[5])
