==> notebooks/10_produce_demo.ipynb <==
import numpy as np
import pandas as pd

from sklearn.externals import joblib

ranking_rfc_fe = pd.read_csv('data/ranking_rfc_fe.csv')
ranking_rfc_fe.sort_values(by='kendall_tau', inplace=True)

X_val = np.load('data/X_val.npy')
df_cols = np.load('data/df_cols.npy')
X_val = pd.DataFrame(X_val, columns=df_cols)
y_val = np.load('data/y_val.npy')

model_features = np.load('data/rfc_fe_list.npy')
model = joblib.load('data/rnd_forest_best.pk1')

def rank_reviews_by_product_id(X_val, y_val,  model, model_features, product_id):
    """
    Rank the reviews of a Product by their helpfullness given a product id and a model

    Args:
        X_val(pd.DataFrame): product reviews data-set (validation)
        y_val(np.array): the labels of product's helpfullness
        model(sklearn.ensemble.forest.RandomForestClassifier): the model used for classifying the reviews
        model_features(np.array): the features used by the model
        product_id(str): the product id

    Returns:
        pd.DataFrame. Ranked Reviews

    """
    # Query all the reviews of product_id:
    products_reviews = X_val[X_val.ProductId == product_id]
    product_helpfulness_labels = y_val

    # Predict the helpfullness score of the reviews:
    predicted_helpfulness = model.predict(products_reviews[model_features])

    products_reviews['Predicted_Helpfulness'] = pd.Series(predicted_helpfulness, index=products_reviews.index)
    products_reviews = products_reviews.sort_values(
        by=['Predicted_Helpfulness', 'Score', 'Total_Reviews_by_Reviewer', 'WordsCount'], ascending=False)
    
    display_colums = ['Predicted_Helpfulness', 'Summary', 'Text', 'Score']
    

    return products_reviews[display_colums]


## Chosen Examples

# B001EQ56T4 - 0
# B004HLCKMY - 1
# B002Y2OWOC - 1

rank_reviews_by_product_id(X_val, y_val, model, model_features, 'B004HLCKMY')

ranking_rfc_fe.sort_values(by='kendall_tau', ascending=False)

type(model)
